+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Basic multi-cluster support for LSF by specifying name of cluster for -m
   argument. [#24](https://github.com/OSC/ood_core/issues/24)
+- Add `header` and `footer` options to a Batch Connect template.
+  [#64](https://github.com/OSC/ood_core/issues/64)
 
 ### Fixed
 - Replaced `Fixnum` code comments with `Integer`.